Add PRNG and Random (#3)
This PR introduces a protocol named `PRNG` for pseudo-random number generators, which is class-constrained and refines `IteratorProtocol` and `Sequence`. Extension methods on `PRNG` should not be susceptible to "modulo bias" problems that can occur when `Element.max - Element.min + 1` is not a power of two.

It also introduces a (final) class named `Random`, which conforms to `PRNG` and implements the `xorshift128+` algorithm, and another (final) class named `Random.Xoroshiro`, which conforms to `PRNG` and implements the `xoroshiro128+` algorithm. Much work remains to be done, including testing the code using [TestU01](http://simul.iro.umontreal.ca/testu01/tu01.html).


* Initial implementation of Random

* Implement discrete uniform distribution

* Make minor change to Random.init?()

* Create protocol PRNG and make changes to Random accordingly

* Revert "Modify FloatingPointMath to refine BinaryFloatingPoint"

* Update README.md

* Simplify some implementations

* Update README.md

* Make minor stylistic changes

* Update implementation of Random.init?()

* Implement MT19937 and xoroshiro128+

* Improve implementation of PRNG.uniform<T : BinaryFloatingPoint>()

* Implement PRNG.exponential and PRNG.weibull

* Refine implementation of Random.MersenneTwister

* Revise description of Xoroshiro128+

* Implement PRNG.bernoulli functions

* Add tests for xorshift128+ and xoroshiro128+, fix and update implementations, make minor stylistic changes

* Change visibility of MT19937 parameters

* Remove Random.MersenneTwister, avoid undefined behavior in PRNG._entropy

* Update to avoid heterogeneous comparison

* Revert "Update to avoid heterogeneous comparison"

* Document methods on PRNG

* Update README.md